// types/ateliere-live.d.ts
export interface ResourcesHTMLBrowser {
  /**
   * Height in pixels of the browser canvas
   * @min 20
   * @max 8192
   * @example 720
   */
  height: number;
  /**
   * The input slot this HTML browser is allocated to
   * @example 1
   */
  input_slot: number;
  /**
   * The active URL in the browser
   * @example "https://www.example.com"
   */
  url: string;
  /**
   * Width in pixels of the browser canvas
   * @min 20
   * @max 8192
   * @example 1280
   */
  width: number;
}
// ...
export interface ResourcesMediaPlayer {
  /**
   * The filename/path of the file to play
   * @example "/media/news.mp4"
   */
  filename: string;
  /**
   * The input slot this media player is allocated to
   * @example 2
   */
  input_slot: number;
}
// ...
export interface ResourcesRenderingEngineResponse {
  html: ResourcesHTMLBrowser[];
  media: ResourcesMediaPlayer[];
}